+++ package.json
@@ -15,13 +15,15 @@
  "description": "Replace this description with project's",
  "devDependencies": {
   "@tailwindcss/typography": "^0.4.0",
-  "lazysizes": "^5.3.0",
+  "jquery": "^3.6.0",
+  "lazysizes": "^5.3.2",
   "npm-run-all": "^4.1.5",
   "postcss-cli": "^8.0.0",
   "postcss-import": "^12.0.1",
   "quicklink": "^2.0.0",
   "react": "^16.14.0",
   "react-dom": "^16.14.0",
+  "scrollmonitor": "^1.2.4",
   "tailwindcss": "^2.0.2"
  },
  "hugo": {
@@ -53,5 +55,6 @@
   "start": "npm run dev:hugo",
   "version": "auto-changelog -p --template keepachangelog --unreleased true"
  },
- "version": "0.1.0"
+ "version": "0.1.0",
+ "dependencies": {}
 }
